Refactor navbar buttons
