+++ src/Footer.js
@@ -1,68 +1,12 @@
 import {
-  Box, Button, Container, Typography,
+  Box, Container, Typography,
 } from '@mui/material';
 import Link from 'next/link';
-import PropTypes from 'prop-types';
-import { socialLinks } from './social-links';
+import NavButton from './NavButton';
+import { leftNav, socialLinks } from './NavButton/social-links';
 import {
   entranceAnimationDelay, entranceAnimationDuration, fadeIn, fadeInLogo,
 } from './utils/animations';
-
-function NavButton({
-  text, href, icon, index,
-}) {
-  return (
-    <Link href={href} passHref>
-      <Button
-        variant="text"
-        sx={{
-          color: 'text.primary',
-          minWidth: icon ? '1rem' : 'inherit',
-          mx: '0.5rem',
-          fontSize: '1.5rem',
-          lineHeight: '2rem',
-          letterSpacing: '0.1rem',
-          animation: `${fadeIn} ${entranceAnimationDuration}s both ${entranceAnimationDelay + index * 0.2}s`,
-        }}
-        target={(icon || href[0] !== '/') ? '_blank' : ''}
-      >
-        {text}
-        {icon || ''}
-      </Button>
-    </Link>
-  );
-}
-
-NavButton.propTypes = {
-  text: PropTypes.string,
-  href: PropTypes.string,
-  icon: PropTypes.node,
-  index: PropTypes.number.isRequired,
-};
-
-NavButton.defaultProps = {
-  text: '',
-  href: '',
-  icon: null,
-};
-
-const leftNav = [
-  {
-    text: 'Home',
-    href: '/',
-    icon: '',
-  },
-  {
-    text: 'Projects',
-    href: '/#projects',
-    icon: '',
-  },
-  {
-    text: 'Contact',
-    href: '/#contact',
-    icon: '',
-  },
-];
 
 export default function Footer() {
   return (
