+++ src/Home/index.js
@@ -1,4 +1,5 @@
 import { Container } from '@mui/material';
+import Contact from '../Contact';
 import OtherWork from '../OtherWork';
 import Projects from '../Projects';
 import Hero from './Hero';
