+++ src/NavButton/index.js
@@ -1,0 +1,44 @@
+import { Button } from '@mui/material';
+import Link from 'next/link';
+import PropTypes from 'prop-types';
+import {
+  entranceAnimationDelay, entranceAnimationDuration, fadeIn,
+} from '../utils/animations';
+
+export default function NavButton({
+  text, href, icon, index,
+}) {
+  return (
+    <Link href={href} passHref>
+      <Button
+        variant="text"
+        sx={{
+          color: 'text.primary',
+          minWidth: icon ? '1rem' : 'inherit',
+          mx: '0.5rem',
+          fontSize: '1.5rem',
+          lineHeight: '2rem',
+          letterSpacing: '0.1rem',
+          animation: `${fadeIn} ${entranceAnimationDuration}s both ${entranceAnimationDelay + index * 0.2}s`,
+        }}
+        target={(icon || href[0] !== '/') ? '_blank' : ''}
+      >
+        {text}
+        {icon || ''}
+      </Button>
+    </Link>
+  );
+}
+
+NavButton.propTypes = {
+  text: PropTypes.string,
+  href: PropTypes.string,
+  icon: PropTypes.node,
+  index: PropTypes.number.isRequired,
+};
+
+NavButton.defaultProps = {
+  text: '',
+  href: '',
+  icon: null,
+};
